+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,32 +1,13 @@
 package roomescape.repository;
 
-import java.util.ArrayList;
 import java.util.List;
 import roomescape.domain.Reservation;
 
-public class ReservationRepository {
+public interface ReservationRepository {
 
-    private final List<Reservation> reservations;
+    List<Reservation> findAll();
 
-    public ReservationRepository() {
-        this.reservations = new ArrayList<>();
-    }
+    void save(final Reservation reservation);
 
-
-    public List<Reservation> findAll() {
-        return reservations;
-    }
-
-    public void save(final Reservation reservation) {
-        reservations.add(reservation);
-    }
-
-    public void delete(final Long id) {
-        Reservation reservation = reservations.stream()
-                .filter(it -> it.getId().equals(id))
-                .findFirst()
-                .orElseThrow(IllegalArgumentException::new);
-
-        reservations.remove(reservation);
-    }
+    void delete(final Long id);
 }
